+++ Assets/curif/LibRetroWrapper/basic/functions/cabsdb.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a5f02cefda0b0fc4eb4bdcbac00b03ed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
